[ADDED] Endpoint returning an account by its number. [DELETED] Endpoint returning an agent by its account number
