+++ app/Http/Controllers/AccountsController.php
@@ -4,6 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Account;
 use App\Transformers\AccountsTranformer;
+use Dingo\Api\Http\Request;
+use Exception;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Validator;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 
 class AccountsController extends AuthController
 {
@@ -53,5 +58,58 @@ class AccountsController extends AuthController
             }
         }
         return $this->response->errorForbidden("Cannot unlink an account that does not belongs to you");
+    }
+
+    /**
+     * Shows an account by its number
+     * @param Request $request
+     * @POST('/api/accounts/by_number')
+     * @Request({account="foo"})
+     */
+    public function show(Request $request) {
+
+        try {
+            $rules = [
+                'account'   => 'required|max:255',
+            ];
+
+            $v = Validator::make($request->all(), $rules);
+            if ($v->fails()) {
+                throw new BadRequestHttpException($v->getMessageBag()->first());
+            }
+            $account = Account::where('number', $request->get('account'))->first();
+            if (!$account) {
+                throw new ModelNotFoundException();
+            }
+         
+            $prefix = $phone = '';
+            if ($account->phone !== null) {
+                $phone_values = explode('-', $agent->phone);
+                $prefix = array_values($phone_values)[0];
+                $phone = array_values($phone_values)[1];
+            }
+            return [
+                'id'                => $account->id,
+                'account'           => $account->number,
+                'owner'             => true,
+                'name'              => $account->alias,
+                'email'             => $account->user->email,
+                'country'           => $account->user->country,
+                'prefix'            => $prefix,
+                'phone'             => $phone,
+                'user_id'           => $account->user->id
+            ];
+            //return $this->item($agent, new AccountsTranformer);
+
+        } catch(BadRequestHttpException $e) {
+            return $this->response()->errorBadRequest($e->getMessage());
+        } catch (ModelNotFoundException $e) {
+            return $this->response()->errorNotFound($e->getMessage());
+        }
+            // @codeCoverageIgnoreStart
+        catch(Exception $e) {
+            return $this->response()->errorInternal();
+        }
+        // @codeCoverageIgnoreEnd
     }
 }
